Move parseCategories to a selector
